refactor reference panel into hooks and mulitple component
